# docs/swagger.yaml
basePath: /
definitions:
  api.CreateJiraConnectionRequest:
    properties:
      authenticationType:
        type: string
      credential:
        type: string
      jiraUrl:
        type: string
      name:
        type: string
      projectKey:
        type: string
      username:
        type: string
    required:
    - authenticationType
    - credential
    - jiraUrl
    - name
    - projectKey
    type: object
  api.JiraConnectionResponse:
    properties:
      authenticationType:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      jiraUrl:
        type: string
      lastTestedAt:
        type: string
      name:
        type: string
      projectId:
        type: string
      projectKey:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  api.UpdateJiraConnectionRequest:
    properties:
      jiraUrl:
        type: string
      name:
        type: string
      projectKey:
        type: string
    type: object
  api.UpdateJiraCredentialsRequest:
    properties:
      authenticationType:
        type: string
      credential:
        type: string
      username:
        type: string
    required:
    - authenticationType
    - credential
    type: object
info:
  contact:
    name: Fern Platform Team
    url: https://github.com/guidewire-oss/fern-platform/issues
  description: Unified test intelligence platform — aggregate, analyze and act on
    test results from any CI/CD pipeline.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/guidewire-oss/fern-platform
  title: Fern Platform API
  version: "1.0"
paths:
...
  /api/v1/projects/{projectId}/integrations/jira/connections:
    get:
      description: Returns all JIRA connections configured for a project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.JiraConnectionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List JIRA connections
      tags:
      - jira
    post:
      consumes:
      - application/json
      description: Configures a JIRA integration for a project (manager or admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: JIRA connection details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateJiraConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.JiraConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a JIRA connection
      tags:
      - jira
  /api/v1/projects/{projectId}/integrations/jira/connections/{connectionId}:
    delete:
      description: Permanently removes a JIRA connection from a project (manager or
        admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a JIRA connection
      tags:
      - jira
    put:
      consumes:
      - application/json
      description: Updates name, URL, or project key for a JIRA connection (manager
        or admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateJiraConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JiraConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a JIRA connection
      tags:
      - jira
  /api/v1/projects/{projectId}/integrations/jira/connections/{connectionId}/credentials:
    put:
      consumes:
      - application/json
      description: Replaces the authentication type and credential for a JIRA connection
        (manager or admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: string
      - description: New credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateJiraCredentialsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JiraConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update JIRA connection credentials
      tags:
      - jira
  /api/v1/projects/{projectId}/integrations/jira/connections/{connectionId}/test:
    post:
      description: Verifies that the stored credentials can reach JIRA (manager or
        admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Test a JIRA connection
      tags:
      - jira
...
securityDefinitions:
  BearerAuth:
    description: 'JWT bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...